requirements.txt
altgraph==0.16.1
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cycler==0.10.0
func-timeout==4.3.0
future==0.17.1
imageio==2.5.0
jieba==0.39
kiwisolver==1.0.1
lxml==4.3.2
macholib==1.11
matplotlib==3.0.3
numpy==1.16.2
pdfminer.six==20181108
pefile==2018.8.8
Pillow==5.4.1
pycryptodome==3.7.3
PyInstaller==3.4
pyparsing==2.3.1
PyQt5==5.11.3
PyQt5-sip==4.19.14
pyqt5-tools==5.11.3.1.4
python-dateutil==2.8.0
python-docx==0.8.10
python-dotenv==0.10.1
pywin32==224
pywin32-ctypes==0.2.0
QScintilla==2.11.1
scipy==1.2.1
six==1.12.0
sortedcontainers==2.1.0
wincertstore==0.2
wordcloud==1.5.0
xlrd==1.2.0
